+++ man-roxygen/options.R
@@ -14,6 +14,13 @@
 #' These global option changes the style of the default column headers:
 #'
 #' * `options(modelsummary_model_labels = "roman")`
+#' * `options(modelsummary_model_labels_term = "Parameter")`
+#' * `options(modelsummary_model_labels_group = "Group")`
+#' * `options(modelsummary_model_labels_model = "Model")`
+#'
+#' The options above control the label shown in the stub columns for terms,
+#' groups, and model names. By default, those headers are blank. Setting any of
+#' these options to a character string inserts that label as the column header.
 #'
 #' The supported styles are: "model", "arabic", "letters", "roman", "(arabic)", "(letters)", "(roman)"
 #'